safari-534.53-branch/LayoutTests/fast/forms/get-forms-to-about-blank-expected.txt

Submitting a GET form to <about:blank> should redirect the browser to <about:blank?>.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS iframe.contentDocument.URL is "about:blank"
PASS iframe.contentDocument.URL is "about:blank?"
PASS successfullyParsed is true

TEST COMPLETE
